optimize

+++ include/utils_c/checksum.h
@@ -16,10 +16,10 @@ extern "C" {
 const static uint16_t OC_CHECKSUM_INIT = 0;
 
 uint16_t
-gen_oc_checksum(uint16_t checksum, const uint8_t *data, size_t len);
+gen_oc_checksum(uint32_t checksum, const uint8_t *data, size_t len);
 
 static inline uint16_t
-finalize_oc_checksum(uint16_t checksum)
+finalize_oc_checksum(uint32_t checksum)
 {
   return (uint16_t)(~checksum & 0xffff);
 }
